+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [0.2.1] - 2019-01-31
+### Added
+- The `Topography` class has been added.
+- `Transform.OfPoint(Vector3 vector)` has been added to transform a vector as a point with translation. This was previously `Transform.OfVector(Vector3 vector)`. All sites previously using `OfVector(...)` are now using `OfPoint(...)`.
+- `Material.DoubleSided`
+- `Loop.InsertEdgeAfter()`
+- `Solid.Slice()`
+### Changed
+- `Transform.OfVector(Vector3 vector)` now does proper vector transformation without translation.
+- Attempting to construct a `Vector3` with NaN or Infinite arguments will throw an `ArgumentOutOfRangeException`.
+
+
 ## [0.2.0] - 2019-01-31
 ### Added
 - IFC implementation has begun with `Model.FromIFC(...)`. Support includes reading of Walls, Slabs, Spaces, Beams, and Columns. Brep booleans required for Wall and Slab openings are not yet supported and are instead converted to Polygon openings in Wall and Floor profiles.
@@ -17,3 +29,4 @@
 - The `Extrusion` class and `IBrep` have been replaced with the `Solid` class. The IGeometry interface now returns a `Solid[]`.
 - Many uses of `System.Linq` have been removed. 
 - Many uses of `IEnumerable<T>` have been replaced with `T[]`.
+